requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.2.1
httpx==0.27.0
alembic==1.13.1
psycopg2-binary==2.9.9
sqlglot==26.0.0
sqlfluff==3.2.0
itsdangerous==2.1.2
PyYAML==6.0.2
beautifulsoup4==4.12.3
lxml==5.2.2
trafilatura==1.12.2
readability-lxml==0.8.1
markdownify==0.11.6
